compress proof in async command
