+++ packages/core/minikit.ts
@@ -1,11 +1,11 @@
 import { VerificationLevel } from '@worldcoin/idkit-core';
 import { encodeAction, generateSignal } from '@worldcoin/idkit-core/hashing';
 import { validatePaymentPayload } from 'helpers/payment/client';
+import { compressAndPadProof } from 'helpers/proof';
 import { generateSiweMessage } from 'helpers/siwe/siwe';
 import { validateWalletAuthCommandInput } from 'helpers/siwe/validate-wallet-auth-command-input';
 import { validateSendTransactionPayload } from 'helpers/transaction/validate-payload';
 import { getUserProfile } from 'helpers/usernames';
-import { compressProof } from 'semaphore-rs-bundler';
 import {
   AsyncHandlerReturn,
   Command,
@@ -33,7 +33,6 @@ import {
   MiniKitInstallErrorMessage,
 } from 'types/errors';
 import { Network } from 'types/payment';
-import { decodeAbiParameters, encodeAbiParameters } from 'viem';
 import { sendWebviewEvent } from './helpers/send-webview-event';
 import {
   EventHandler,
@@ -137,41 +136,7 @@ export class MiniKit {
         payload,
       ) => {
         if (payload.status === 'success') {
-          // Decode the hex proof to array of 8 uints
-          const decodedProof = decodeAbiParameters(
-            [{ type: 'uint256[8]' }],
-            payload.proof as `0x${string}`,
-          )[0] as readonly [
-            bigint,
-            bigint,
-            bigint,
-            bigint,
-            bigint,
-            bigint,
-            bigint,
-            bigint,
-          ];
-
-          // Convert to hex strings for compression
-          const proofHexStrings = [...decodedProof].map(
-            (p) => '0x' + p.toString(16).padStart(64, '0'),
-          ) as [string, string, string, string, string, string, string, string];
-
-          // Compress and pad the proof
-          const compressedProof = compressProof(proofHexStrings);
-          const paddedProof = [
-            ...compressedProof.map(BigInt),
-            0n,
-            0n,
-            0n,
-            0n,
-          ] as [bigint, bigint, bigint, bigint, bigint, bigint, bigint, bigint];
-
-          // Encode back to hex string
-          payload.proof = encodeAbiParameters(
-            [{ type: 'uint256[8]' }],
-            [paddedProof],
-          );
+          payload.proof = compressAndPadProof(payload.proof as `0x${string}`);
         }
         originalHandler(payload);
       };
@@ -578,6 +543,11 @@ export class MiniKit {
             Command.Verify,
             () => this.commands.verify(payload),
           );
+          if (response.finalPayload.status === 'success') {
+            response.finalPayload.proof = compressAndPadProof(
+              response.finalPayload.proof as `0x${string}`,
+            );
+          }
           resolve(response);
         } catch (error) {
           reject(error);
